+++ R/SNAhelper.R
@@ -21,9 +21,9 @@ SNAhelper <- function(text){
   if (!requireNamespace("graphlayouts", quietly = TRUE)) {
     stop("graphlayouts package not found. Install it with install.packages('graphlayouts')", call. = FALSE)
   }
-  if(utils::packageVersion("ggraph")<"1.0.2.9999"){
-    stop("snahelper needs a newer version of ggraph. please update it")
-  }
+  # if(utils::packageVersion("ggraph")<"1.0.2.9999"){
+  #   stop("snahelper needs a newer version of ggraph. please update it")
+  # }
 
   if (any(ls(envir = .GlobalEnv) == text)) {
     g <- get(text, envir = .GlobalEnv)