+++ NEWS.md
@@ -3,6 +3,7 @@
 * **BREAKING** only works with the newest ggraph version (error message added)
 * edge geom is chosen automatically between `geom_edge_link0()` and `geom_edge_parallel0()`
 * discrete edge colors are supported now
+* added support for `layout_with_focus()` and `layout_with_centrality()`
 
 # snahelper 0.3.0